Trait exolvl::Read

pub trait Read: Sealed {
    // Required method
    fn read(input: &mut impl Read) -> Result<Self, Error>
       where Self: Sized;
}

A trait for reading values from a binary exolvl file.

Sealed

This trait is sealed and cannot be implemented for types outside of this crate.

Required Methods

fn read(input: &mut impl Read) -> Result<Self, Error>
where Self: Sized,

Reads a value from the given input.

Errors

Returns an error if the underlying reader returns an error.