Trait exolvl::ReadContext

pub trait ReadContext: Sealed {
    type Context;

    // Required method
    fn read_ctx(
        input: &mut impl Read,
        with: Self::Context,
    ) -> Result<Self, Error>
       where Self: Sized;
}

A trait for reading values from a binary exolvl file with additional context.

This trait is similar to Read, but takes an additional generic parameter. This is sometimes necessary for reading enums.

Sealed

This trait is sealed and cannot be implemented for types outside of this crate.

Required Associated Types

type Context

Additional context to pass to read_ctx.

Required Methods

fn read_ctx(input: &mut impl Read, with: Self::Context) -> Result<Self, Error>
where Self: Sized,

Reads a value from a given output. This method takes an additional parameter as context.

This method should almost never be called by outside users of the library. Use Read instead.

Errors

Returns an error if the underlying reader returns an error.