exolvl
Trait ReadVersioned

pub trait ReadVersioned: Sealed {
    // Required method
    fn read(input: &mut impl Read, version: i32) -> Result<Self, Error>
       where Self: Sized;
}

A trait to allow reading values from a binary exolvl file while adhering to a specific version of the format.

This allows reading old files while also supporting new features.

Sealed

This trait is sealed and cannot be implemented for types outside of this crate.

Required Methods

fn read(input: &mut impl Read, version: i32) -> Result<Self, Error>
where Self: Sized,

Reads a value from a given output. This method takes an additional parameter of type i32 which describes the level format version.

This method should almost never be called by outside users of the library. Use Read instead. If you need to use this method, take special care to pass the correct version for the specific level file you’re using.

Errors

Returns an error if the underlying reader returns an error.