Trait exolvl::Write

pub trait Write: Sealed {
    // Required method
    fn write(&self, output: &mut impl Write) -> Result<(), Error>;
}

A trait for writing values to a binary exolvl file.

Sealed

This trait is sealed and cannot be implemented for types outside of this crate.

Required Methods

fn write(&self, output: &mut impl Write) -> Result<(), Error>

Writes a value to a given output.

Errors

Returns an error if the underlying writer returns an error.

Object Safety

This trait is not object safe.