Struct exolvl::types::level_data::LevelData

pub struct LevelData {

    pub level_id: Uuid,
    pub level_version: i32,
    pub nova_level: bool,
    pub under_decoration_tiles: Vec<i32>,
    pub background_decoration_tiles: Vec<i32>,
    pub terrain_tiles: Vec<i32>,
    pub floating_zone_tiles: Vec<i32>,
    pub object_tiles: Vec<i32>,
    pub foreground_decoration_tiles: Vec<i32>,
    pub objects: Vec<Object>,
    pub layers: Vec<Layer>,
    pub prefabs: Vec<Prefab>,
    pub brushes: Vec<Brush>,
    pub patterns: Vec<Pattern>,
    pub color_palette: Option<Vec<Color>>,
    pub author_time: i64,
    pub author_lap_times: Vec<i64>,
    pub silver_medal_time: i64,
    pub gold_medal_time: i64,
    pub laps: i32,
    pub center_camera: bool,
    pub scripts: Vec<i32>,
    pub nova_scripts: Vec<NovaScript>,
    pub global_variables: Vec<Variable>,
    pub theme: Theme,
    pub custom_background_color: Color,
    pub custom_terrain_pattern_id: i32,
    pub custom_terrain_pattern_tiling: Vec2,
    pub custom_terrain_pattern_offset: Vec2,
    pub custom_terrain_color: Color,
    pub custom_terrain_secondary_color: Color,
    pub custom_terrain_blend_mode: i32,
    pub custom_terrain_border_color: Color,
    pub custom_terrain_border_thickness: OrderedFloat<f32>,
    pub custom_terrain_border_corner_radius: OrderedFloat<f32>,
    pub custom_terrain_round_reflex_angles: bool,
    pub custom_terrain_round_collider: bool,
    pub custom_terrain_friction: OrderedFloat<f32>,
    pub default_music: bool,
    pub music_ids: Vec<String>,
    pub allow_direction_change: bool,
    pub disable_replays: bool,
    pub disable_revive_pads: bool,
    pub disable_start_animation: bool,
    pub gravity: Vec2,
    /* private fields */

}

The level data for an Exoracer level.

Fields
level_id: Uuid

The UUID of the level.

level_version: i32

The version of the level e.g. v1, v2, etc.

nova_level: bool

Whether this level is for the new level editor.

If this is true, the level can be opened in the new level editor. Otherwise it’s for the “legacy” editor. This Field is presumably only useful in .level files, not in .exolvl ones. A mismatch with the corresponding LocalLevel field should be avoided.

under_decoration_tiles: Vec<i32>

The tile ids for the “under decoration” layer.

background_decoration_tiles: Vec<i32>

The tile ids for the “background decoration” layer.

terrain_tiles: Vec<i32>

The tile ids for the terrain layer.

floating_zone_tiles: Vec<i32>

The tile ids for the floating zone layer.

object_tiles: Vec<i32>

The tile ids for the “object” layer.

foreground_decoration_tiles: Vec<i32>

The tile ids for the “foreground decoration” layer.

objects: Vec<Object>

The objects in the level.

layers: Vec<Layer>

The layers in the level.

prefabs: Vec<Prefab>

The prefabs in the level.

brushes: Vec<Brush>

The brushes in the level.

patterns: Vec<Pattern>

The patterns in the level.

color_palette: Option<Vec<Color>>

The color palettes in the level.

This is only present in levels with version 17 or higher.

author_time: i64

The author medal time for this level in milliseconds.

author_lap_times: Vec<i64>

The author medal lap times for this level in milliseconds.

silver_medal_time: i64

The silver medal time for this level in milliseconds.

gold_medal_time: i64

The gold medal time for this level in milliseconds.

laps: i32

The number of laps in this level.

center_camera: bool

Whether the camera should be centered while playing this level.

This is mostly deprecated and should stay set to false.

scripts: Vec<i32>

The scripts in the level.

These are used in the legacy level editor.

nova_scripts: Vec<NovaScript>

The “new” scripts in the level.

These are the scripts that are used in the new level editor. As opposed to the scripts field which is for the legacy editor.

global_variables: Vec<Variable>

All the global variables in the level.

theme: Theme

The theme name of the level.

custom_background_color: Color

The custom background color of the level.

custom_terrain_pattern_id: i32

The following terrain related fields are all used when explicitly copying certain terrain data.

The custom terrain pattern that can be pasted with the color_paste button if the recieving object has the FillMode set to Pattern.

custom_terrain_pattern_tiling: Vec2

The tiling of that pattern.

custom_terrain_pattern_offset: Vec2

the offset of that pattern.

custom_terrain_color: Color

In the legacy editor: The custom terrain color of the level. In the new editor: The color of the copied terrain.

custom_terrain_secondary_color: Color

Not 100% sure of the use of this, presumably the replacement for the border color in the new editor. Used when copying and pasting properties of terrain.

custom_terrain_blend_mode: i32

The blend mode of the copied terrain.

custom_terrain_border_color: Color

The custom terrain border color of the level.

custom_terrain_border_thickness: OrderedFloat<f32>

The thickness of the terrain border.

custom_terrain_border_corner_radius: OrderedFloat<f32>

The corner radius of the terrain border.

custom_terrain_round_reflex_angles: bool

Whether the copied terrain has round reflex angles or not (only visual).

custom_terrain_round_collider: bool

Whether the copied terrain has a round collider or not (not visual).

custom_terrain_friction: OrderedFloat<f32>

The friction of the copied terrain.

default_music: bool

Whether the default music should be played or not.

music_ids: Vec<String>

The music ids for the level. The game randomly picks one of these to play each time.

allow_direction_change: bool

Whether the game lets the player change directions or not.

disable_replays: bool

Whether replays are disabled or not.

If this is true, the game won’t upload replays on this level. (unless you explicitly upload blank shells from the history section, that only contain the time you set without any replay data. Could be a bug).

disable_revive_pads: bool

Whether revive pads are disabled or not.

disable_start_animation: bool

Whether the start animation is disabled or not.

gravity: Vec2

The gravity vector for this level.

Implementations

impl LevelData

pub fn default_with_id(level_id: Uuid) -> Self